:recycle: make arrayDeepFilter available through a trait
